refactor(client/clientimplementation): split functions across file to reduce file size
